requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
PyMuPDF
pytesseract
pandas
elasticsearch
pdfminer.six
python-dotenv
tqdm
